perf: resolve all performance audit findings
Schedule Generator Engine:
- Reuse game objects in slot allocator (eliminate duplicate create_game per slot)
- Batch cancellation/timeout checks every 25 matchups (96% fewer DB hits)
- Add matchup count hash map for O(1) lookups (was O(N) linear scan)
- Batch track_changes() into single update_option() call
- Batch importer progress transient updates every 25 games
- Cache team/venue name lookups in importer (O(1) vs O(N) per game)
- Lazy-init config manager and renderer (only on schedule generator page)
- Remove duplicate nonces between spsgData and spsgAdminData
- Lazy-load configuration (remove init hook, load on first access)
- Cast max_generation_time to int for type safety

Events Manager:
- Batch calendar lookup in create_missing_calendars() (eliminate N+1)
- Prime meta/term caches in reset_calendars_to_current_season()
- Add instance caches for find_or_create_team/venue/league
- Cache performance_keys and result_keys during import

e-Transfer Automation:
- Cache pending webhook count per request
- Single log fetch shared between display methods
- Pre-fetch WooCommerce orders before unmatched webhooks loop
- Add summary mode to get_etransfer_logs() (exclude longtext columns)

Player Tools:
- Replace admin_init cleanup with daily cron event
- Replace similar_text() O(N³) with levenshtein() O(N²) in fuzzy matching
- Pre-fetch team/player titles in single query
- Cache player posts per user in profile picture handler
- Batch bulk_enable_stats() with 100-player pages

+++ sportspress-schedule-generator/includes/class-schedule-engine.php
@@ -96,7 +96,7 @@ class SPSG_Schedule_Engine
         $this->init_progress_tracking();
 
         // Get max generation time from config or use default
-        $this->max_generation_time = get_option('spsg_max_generation_time', 300);
+        $this->max_generation_time = (int) get_option('spsg_max_generation_time', 300);
 
         // Check for cancellation before starting
         if ($this->is_cancelled()) {
